+++ types/Core/ASelector.ts
