+++ types/Core/Base.ts
