+++ types/Core/Command/CommandCooldown.ts
