+++ types/Core/Command/CommandPermissions.ts
