+++ types/Core/CommandDispatcher.ts
