+++ types/Core/Models/GuildConfig.ts
