+++ types/Core/Stores/ARegistry.ts
