+++ types/Core/Stores/CommandRegistry.ts
