+++ types/Core/Stores/GuildConfigCache.ts
