+++ types/Core/Stores/ModuleRegistry.ts
